Commit message: "Add fruit editing functionality"
This commit message summarizes that you have added the fruit editing functionality to the project. The code in this commit includes an editController function that handles editing a fruit's details. It uses the editById and getById functions from the apiFruits.js module to retrieve the fruit's data and update it after editing.