Commit message: "Add fruit details page and functionality"
This commit message summarizes that you have added the fruit details page and associated functionality to the project. The code in this commit includes a detailsController function that handles displaying the details of a fruit. It uses the getById function from the apiFruits.js module to retrieve the fruit's data. The template displays the fruit details, including the image, description, and nutrition information. If the user is the creator of the fruit, the template also includes "Edit" and "Delete" buttons.