Commit message: "Implement fruit catalog page"
This commit message summarizes that you have implemented the fruit catalog page and associated functionality in the project. The code in this commit includes a catalogController function that handles displaying the catalog of fruits. It uses the getAll function from the apiFruits.js module to retrieve all fruits' data. The template displays a list of fruits with their images, names, descriptions, and "More Info" links.